requirements/prod.txt
openai
pydub
imagehash
pillow